[refactor] move annotation to new directory
