+++ src/main/java/scs/comp5903/cucumber/parser/JStepDefinitionParser.java
@@ -4,7 +4,7 @@ import org.slf4j.Logger;
 import scs.comp5903.cucumber.model.JStepDefDetail;
 import scs.comp5903.cucumber.model.JStepDefMethodDetail;
 import scs.comp5903.cucumber.model.JStepKeyword;
-import scs.comp5903.cucumber.model.annotation.*;
+import scs.comp5903.cucumber.model.annotation.step.*;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
 import scs.comp5903.cucumber.model.exception.ErrorCode;
 import scs.comp5903.cucumber.model.matcher.*;
