+++ src/test/java/scs/comp5903/cucumber/sample/CmsSimpleRegisterCourseStepDefs.java
@@ -3,7 +3,7 @@ package scs.comp5903.cucumber.sample;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.slf4j.Logger;
-import scs.comp5903.cucumber.model.annotation.JStep;
+import scs.comp5903.cucumber.model.annotation.step.JStep;
 import scs.comp5903.cucumber.sample.cmsutil.CmsPageUtils;
 import scs.comp5903.cucumber.sample.cmsutil.Constants;
 import scs.comp5903.cucumber.sample.cmsutil.SeleniumFactory;
