+++ src/test/java/scs/comp5903/cucumber/sample/DecimalAndBigIntStepDef.java
@@ -2,7 +2,7 @@ package scs.comp5903.cucumber.sample;
 
 import org.slf4j.Logger;
 import scs.comp5903.cucumber.model.JStepKeyword;
-import scs.comp5903.cucumber.model.annotation.JStep;
+import scs.comp5903.cucumber.model.annotation.step.JStep;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
