+++ src/test/java/scs/comp5903/cucumber/sample/samplestepdef/RummikubDummySeparatedJStepDefs.java
@@ -1,7 +1,7 @@
 package scs.comp5903.cucumber.sample.samplestepdef;
 
 import org.slf4j.Logger;
-import scs.comp5903.cucumber.model.annotation.JStep;
+import scs.comp5903.cucumber.model.annotation.step.JStep;
 
 import static scs.comp5903.cucumber.model.JStepKeyword.*;
